requirements.txt
ebooklib==0.20
beautifulsoup4==4.14.2
markdownify==1.2.2
pdfminer.six==20251107
pypdf==5.2.0
python-docx==1.2.0
click==8.3.1
